Working with stores
In this example I show how I use stores. I created another file that contained a message variable which says "Hello". I simply access that variable by using {$message} and then using an on:click  which will call a function that changes the value of the message variable to Bye.